Replace double with decimal in registry item model

+++ WeddingWebsite/Models/Registry/RegistryItem.cs
@@ -15,7 +15,7 @@ public record RegistryItem(
 {
     public int QuantityClaimed => Claims.Sum(c => c.Quantity);
     public bool IsFullyClaimed => QuantityClaimed >= MaxQuantity;
-    public double CheapestCost => PurchaseMethods.Min(pm => pm.Cost);
+    public decimal CheapestCost => PurchaseMethods.Min(pm => pm.Cost);
     
     public int NumClaimsByUser(string userId)
     {
